IS-3357: Move dialogmelding-client to infrastructure
