+++ src/main/kotlin/no/nav/syfo/infrastructure/cronjob/CronjobModule.kt
@@ -7,7 +7,7 @@ import no.nav.syfo.infrastructure.client.cache.ValkeyStore
 import no.nav.syfo.infrastructure.database.DatabaseInterface
 import no.nav.syfo.launchBackgroundTask
 import no.nav.syfo.application.ArbeidstakerVarselService
-import no.nav.syfo.client.dialogmelding.DialogmeldingClient
+import no.nav.syfo.infrastructure.client.dialogmelding.DialogmeldingClient
 import no.nav.syfo.infrastructure.client.azuread.AzureAdV2Client
 import no.nav.syfo.infrastructure.client.pdl.PdlClient
 import no.nav.syfo.infrastructure.cronjob.dialogmoteOutdated.DialogmoteOutdatedCronjob
